install: autodetect filesystem before attempting format
* for f2fs/ext4, autodetect filesystem that is already on the
  block and use that for formatting if possible

+++ updater/install.c
@@ -348,33 +348,39 @@ Value* FormatFn(const char* name, State* state, int argc, Expr* argv[]) {
         }
         result = location;
 #ifdef USE_EXT4
-    } else if (strcmp(fs_type, "ext4") == 0) {
-        int status = make_ext4fs(location, atoll(fs_size), mount_point, sehandle);
-        if (status != 0) {
-            printf("%s: make_ext4fs failed (%d) on %s",
-                    name, status, location);
-            result = strdup("");
-            goto done;
+    } else if (strcmp(fs_type, "ext4") == 0 || strcmp(fs_type, "f2fs") == 0) {
+        char *detected_fs_type = blkid_get_tag_value(NULL, "TYPE", rec->blk_device);
+        if (detected_fs_type != NULL && strcmp(fs_type, detected_fs_type) != 0)
+            fs_type = detected_fs_type;
+
+        if (strcmp(fs_type, "ext4") == 0) {
+            int status = make_ext4fs(location, atoll(fs_size), mount_point, sehandle);
+            if (status != 0) {
+                printf("%s: make_ext4fs failed (%d) on %s",
+                        name, status, location);
+                result = strdup("");
+                goto done;
+            }
+            result = location;
+        } else if (strcmp(fs_type, "f2fs") == 0) {
+            char *num_sectors;
+            if (asprintf(&num_sectors, "%lld", atoll(fs_size) / 512) <= 0) {
+                printf("format_volume: failed to create %s command for %s\n", fs_type, location);
+                result = strdup("");
+                goto done;
+            }
+            const char *f2fs_path = "/sbin/mkfs.f2fs";
+            const char* const f2fs_argv[] = {"mkfs.f2fs", "-t", "-d1", location, num_sectors, NULL};
+            int status = exec_cmd(f2fs_path, (char* const*)f2fs_argv);
+            free(num_sectors);
+            if (status != 0) {
+                printf("%s: mkfs.f2fs failed (%d) on %s",
+                        name, status, location);
+                result = strdup("");
+                goto done;
+            }
+            result = location;
         }
-        result = location;
-    } else if (strcmp(fs_type, "f2fs") == 0) {
-        char *num_sectors;
-        if (asprintf(&num_sectors, "%lld", atoll(fs_size) / 512) <= 0) {
-            printf("format_volume: failed to create %s command for %s\n", fs_type, location);
-            result = strdup("");
-            goto done;
-        }
-        const char *f2fs_path = "/sbin/mkfs.f2fs";
-        const char* const f2fs_argv[] = {"mkfs.f2fs", "-t", "-d1", location, num_sectors, NULL};
-        int status = exec_cmd(f2fs_path, (char* const*)f2fs_argv);
-        free(num_sectors);
-        if (status != 0) {
-            printf("%s: mkfs.f2fs failed (%d) on %s",
-                    name, status, location);
-            result = strdup("");
-            goto done;
-        }
-        result = location;
 #endif
     } else {
         printf("%s: unsupported fs_type \"%s\" partition_type \"%s\"",
